fix(lens): lift windowedSub into a shared leaf so the removal-race floor lives in one place
Move windowedSub out of activeWire.ts into hostScope/windowedSub.ts (pure Subscription-shape helper, zero wire/activeScope deps); import it in activeWire.ts and perHostPolledQuery.ts, replacing perHostPolledQuery's hand-rolled Object.assign facade with windowedSub(() => scopes.active(), (v) => v, undefined).

Agreed by the lowy ⇄ hickey lens debate (finding lowy-1, raised by lowy). Not pushed or merged.
